packet drops are now in %

+++ config.h
@@ -46,7 +46,7 @@ typedef struct {
 	uint64_t max_rate;
 	uint64_t step_min;
 	uint64_t step_max;
-	int drop;
+	float drop;
 
 	int packet_size;
 	int dst_macs_are_set;
